put for Billing Account

+++ fapi/src/main/java/com/netcracker/edu/fapi/controller/SubscriptionsController.java
@@ -18,20 +18,20 @@ public class SubscriptionsController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ResponseEntity<List<SubscriptionViewModel>> getSubscriptionsByUserId(@PathVariable String id) {
-        return ResponseEntity.ok(subscriptionsService.getSubscriptions(Integer.valueOf(id)));
+        return ResponseEntity.ok(subscriptionsService.getSubscriptions(Long.valueOf(id)));
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.POST)
     public ResponseEntity<SubscriptionViewModel> addSubscription(@PathVariable String id, @RequestBody SubscriptionViewModel subscription) {
         if (subscription != null) {
-            return ResponseEntity.ok(subscriptionsService.addSubscription(Integer.valueOf(id), subscription));
+            return ResponseEntity.ok(subscriptionsService.addSubscription(Long.valueOf(id), subscription));
         }
         return ResponseEntity.badRequest().build();
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity deleteSubscription(@PathVariable String id) {
-        subscriptionsService.deleteSubscription(Integer.valueOf(id));
+        subscriptionsService.deleteSubscription(Long.valueOf(id));
         return ResponseEntity.ok().build();
     }
 }
